http, cttp: add header file
