+++ pkg/urbit/include/vere/vere.h
@@ -15,24 +15,6 @@
   /** Data types.
   **/
 
-    /* u3_hhed: http header.
-    */
-      typedef struct _u3_hhed {
-        struct _u3_hhed* nex_u;
-        c3_w             nam_w;
-        c3_c*            nam_c;
-        c3_w             val_w;
-        c3_c*            val_c;
-      } u3_hhed;
-
-    /* u3_hbod: http body block.  Also used for responses.
-    */
-      typedef struct _u3_hbod {
-        struct _u3_hbod* nex_u;
-        c3_w             len_w;
-        c3_y             hun_y[0];
-      } u3_hbod;
-
     /* u3_lane: ames lane (IP address and port)
     */
       typedef struct _u3_lane {
@@ -1114,20 +1096,6 @@
       */
         u3_auto*
         u3_behn_io_init(u3_pier* pir_u);
-
-    /**  HTTP server.
-    **/
-      /* u3_http_io_init(): initialize http I/O.
-      */
-        u3_auto*
-        u3_http_io_init(u3_pier* pir_u);
-
-    /**  HTTP client.
-    **/
-      /* u3_cttp_io_init(): initialize cttp I/O.
-      */
-        u3_auto*
-        u3_cttp_io_init(u3_pier* pir_u);
 
     /**  fore, first events
     **/
